requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.9.2
billiard==4.2.2
boto3==1.40.64
botocore==1.40.64
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.1
dj-database-url==3.0.1
dj-rest-auth==7.0.1
Django==5.2.6
django-allauth==65.11.2
django-cors-headers==4.9.0
django-environ==0.12.0
django-storages==1.14.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gevent==25.9.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.5.0
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
grpcio-status==1.71.2
gunicorn==23.0.0
httplib2==0.31.0
idna==3.10
jmespath==1.0.1
kombu==5.5.4
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==5.29.5
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.5
python-barcode==0.16.0
python-dateutil==2.9.0.post0
qrcode==8.0
redis==6.4.0
reportlab==4.0.9
requests==2.32.5
rsa==4.9.1
s3transfer==0.14.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.8.2
zope.event==6.0
zope.interface==8.0.1
sendgrid==6.11.0
